requirements.txt
autopep8==1.5.7
Brotli==1.0.9
click==8.0.0
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Flask==2.0.0
Flask-Compress==1.9.0
future==0.18.2
gunicorn==20.1.0
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.0.1
llvmlite==0.36.0
MarkupSafe==2.0.0
numba==0.53.1
numpy==1.20.3
pandas==1.2.4
plotly==4.14.3
pycodestyle==2.7.0
pynndescent==0.5.2
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
toml==0.10.2
umap-learn==0.5.1
Werkzeug==2.0.0
